Делаю 1 отправку, с 7 корректировкой

+++ src/task1.cpp
@@ -7,7 +7,7 @@ int comp1(void* a, void* b) {
 }
 
 int comp2(void* a, void* b) {
-	return (int)(*((double*)a) - *((double*)b));
+	return ((int)(*((double*)a)) - (int)(*((double*)b)));
 }
 
 int comp3(void* a, void* b) {
